Add test_transformer_encoder_layer_shape
